refactor(dashboard): move ticker news onto React Query
Replace useTickerNews's hand-rolled module cache + setInterval with a useQuery
that shares the same cache/poll mechanism as useDashboardData, and extract the
duplicated news mapper/formatter/types into a shared newsItem module (fixing the
zh-CN relative-time the watchlist/portfolio path previously missed). getNews now
re-throws on error so the React Query callers retry and keep the last good list
instead of blanking the feed.
