+++ web/src/pages/Dashboard/hooks/useDashboardData.ts
@@ -1,39 +1,22 @@
 import { useQuery, useInfiniteQuery } from '@tanstack/react-query';
 import { useMemo } from 'react';
-import i18n from '@/i18n';
 import { getNews, getIndices, INDEX_SYMBOLS, fallbackIndex, normalizeIndexSymbol } from '../utils/api';
 import { fetchMarketStatus } from '@/lib/marketUtils';
 import type { IndexData } from '@/types/market';
+import {
+  type DashboardNewsItem,
+  NEWS_POLL_INTERVAL_MS,
+  NEWS_STALE_MS,
+  mapNewsResults,
+} from '../utils/newsItem';
+
+type NewsItem = DashboardNewsItem;
 
 interface MarketStatusData {
   market?: string;
   afterHours?: boolean;
   earlyHours?: boolean;
   [key: string]: unknown;
-}
-
-interface NewsSentimentItem {
-  ticker: string;
-  sentiment: string;
-  reasoning?: string;
-}
-
-interface NewsItem {
-  id: string;
-  title: string;
-  time: string;
-  publishedAt: string | null;
-  isHot: boolean;
-  source: string;
-  favicon: string | null;
-  image: string | null;
-  tickers: string[];
-  articleUrl?: string | null;
-  // Inlined article body — lets the detail modal render without a by-id fetch.
-  author?: string | null;
-  description?: string | null;
-  keywords?: string[];
-  sentiments?: NewsSentimentItem[] | null;
 }
 
 interface DashboardData {
@@ -48,48 +31,6 @@ interface DashboardData {
   curatedFetchNextPage: () => void;
   marketStatus: MarketStatusData | null;
   marketStatusRef: { current: MarketStatusData | null };
-}
-
-function mapNewsResults(results: Record<string, unknown>[]): NewsItem[] {
-  return results.map((r) => ({
-    id: r.id as string,
-    title: r.title as string,
-    time: formatRelativeTime(r.published_at as string | null | undefined),
-    publishedAt: (r.published_at as string) || null,
-    isHot: r.has_sentiment as boolean,
-    source: (r.source as Record<string, unknown> | undefined)?.name as string || '',
-    favicon: (r.source as Record<string, unknown> | undefined)?.favicon_url as string || null,
-    image: r.image_url as string || null,
-    tickers: (r.tickers as string[]) || [],
-    articleUrl: (r.article_url as string) || null,
-    author: (r.author as string) ?? null,
-    description: (r.description as string) ?? null,
-    keywords: (r.keywords as string[]) || [],
-    sentiments: (r.sentiments as NewsSentimentItem[]) ?? null,
-  }));
-}
-
-/**
- * Formats a given timestamp to a relative time string. Outside React render —
- * components consuming the result via this hook re-render on locale switch
- * because their parent calls useTranslation, which is what makes the freshly
- * resolved string reach the DOM.
- */
-function formatRelativeTime(timestamp: string | number | null | undefined): string {
-  if (!timestamp) return '';
-  const now = new Date();
-  const then = new Date(timestamp);
-  const diffMs = now.getTime() - then.getTime();
-  const diffMin = Math.floor(diffMs / 60000);
-  if (diffMin < 1) return i18n.t('dashboard.widgets.common.relativeNow');
-  let when: string;
-  if (diffMin < 60) when = `${diffMin}m`;
-  else {
-    const diffHr = Math.floor(diffMin / 60);
-    if (diffHr < 24) when = `${diffHr}h`;
-    else when = `${Math.floor(diffHr / 24)}d`;
-  }
-  return i18n.t('dashboard.widgets.common.relativePast', { when });
 }
 
 /**
@@ -125,23 +66,33 @@ export function useDashboardData(): DashboardData {
     staleTime: 10000,
   });
 
-  // 3. News Feed (Fetched once, cached for 5 minutes)
+  // 3. Market General Feed — kept warm server-side by the news poller, so we
+  //    re-poll every 60s to surface the latest articles in an open tab.
   const { data: newsItems = [], isLoading: newsLoading } = useQuery<NewsItem[]>({
     queryKey: ['dashboard', 'news'],
     queryFn: async (): Promise<NewsItem[]> => {
       const data = await getNews({ limit: 50 });
       return data.results?.length ? mapNewsResults(data.results) : [];
     },
-    staleTime: 5 * 60 * 1000, // 5 minutes fresh cache
+    staleTime: NEWS_STALE_MS,
+    refetchInterval: NEWS_POLL_INTERVAL_MS,
+    refetchIntervalInBackground: false,
   });
 
-  // 4. Curated "Top" Feed (TickerTick) — cursor-paginated for infinite scroll.
+  // 4. Curated "Top" Feed (TickerTick) — cursor-paginated for infinite scroll,
+  //    also kept warm server-side. Auto-refresh ONLY page 1 (the warm buffer):
+  //    refetchInterval refetches every loaded page, and pages 2+ bypass the
+  //    server cache and hit upstream directly, so we stop polling once the user
+  //    scrolls past page 1.
   const curated = useInfiniteQuery({
     queryKey: ['dashboard', 'curatedNews'],
     queryFn: ({ pageParam }) => getNews({ provider: 'tickertick', limit: 50, cursor: pageParam }),
     initialPageParam: undefined as string | undefined,
     getNextPageParam: (lastPage) => lastPage.next_cursor ?? undefined,
-    staleTime: 5 * 60 * 1000, // 5 minutes fresh cache
+    staleTime: NEWS_STALE_MS,
+    refetchInterval: (query) =>
+      (query.state.data?.pages.length ?? 0) <= 1 ? NEWS_POLL_INTERVAL_MS : false,
+    refetchIntervalInBackground: false,
   });
 
   // Flatten loaded pages, de-duping by id (guards against feed rotation between
