+++ web/src/pages/Dashboard/hooks/useTickerNews.ts
@@ -1,120 +1,50 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useQuery } from '@tanstack/react-query';
 import { getNews } from '../utils/api';
+import {
+  type DashboardNewsItem,
+  NEWS_POLL_INTERVAL_MS,
+  NEWS_STALE_MS,
+  mapNewsResults,
+} from '../utils/newsItem';
 
-interface NewsSentimentItem {
-  ticker: string;
-  sentiment: string;
-  reasoning?: string;
-}
-
-export interface TickerNewsItem {
-  id: string;
-  title: string;
-  time: string;
-  publishedAt: string | null;
-  isHot: boolean;
-  image: string | null;
-  source: string;
-  favicon: string | null;
-  tickers: string[];
-  articleUrl?: string | null;
-  // Inlined article body — lets the detail modal render without a by-id fetch.
-  author?: string | null;
-  description?: string | null;
-  keywords?: string[];
-  sentiments?: NewsSentimentItem[] | null;
-}
+// Back-compat alias — the news-item shape is shared across all dashboard feeds.
+export type TickerNewsItem = DashboardNewsItem;
 
 interface TickerRow {
   symbol: string;
   [key: string]: unknown;
 }
 
-interface CacheEntry {
-  items: TickerNewsItem[];
-  tickerKey: string;
-}
-
-// Module-level caches keyed by caller-provided cacheKey
-const cacheMap = new Map<string, CacheEntry>();
-
-function formatRelativeTime(timestamp: string | number | null | undefined): string {
-  if (!timestamp) return '';
-  const now = new Date();
-  const then = new Date(timestamp);
-  const diffMs = now.getTime() - then.getTime();
-  const diffMin = Math.floor(diffMs / 60000);
-  if (diffMin < 1) return 'just now';
-  if (diffMin < 60) return `${diffMin} min ago`;
-  const diffHr = Math.floor(diffMin / 60);
-  if (diffHr < 24) return `${diffHr} hr${diffHr > 1 ? 's' : ''} ago`;
-  const diffDay = Math.floor(diffHr / 24);
-  return `${diffDay} day${diffDay > 1 ? 's' : ''} ago`;
-}
-
-function mapNewsResults(results: Record<string, unknown>[]): TickerNewsItem[] {
-  return results.map((r) => ({
-    id: r.id as string,
-    title: r.title as string,
-    time: formatRelativeTime(r.published_at as string | null | undefined),
-    publishedAt: (r.published_at as string) || null,
-    isHot: r.has_sentiment as boolean,
-    image: r.image_url as string || null,
-    source: (r.source as Record<string, unknown> | undefined)?.name as string || '',
-    favicon: (r.source as Record<string, unknown> | undefined)?.favicon_url as string || null,
-    tickers: (r.tickers as string[]) || [],
-    articleUrl: (r.article_url as string) || null,
-    author: (r.author as string) ?? null,
-    description: (r.description as string) ?? null,
-    keywords: (r.keywords as string[]) || [],
-    sentiments: (r.sentiments as NewsSentimentItem[]) ?? null,
-  }));
-}
-
 /**
- * Hook to fetch news for a list of ticker rows.
+ * Fetches news for a list of ticker rows via React Query — same cache/poll
+ * mechanism as useDashboardData (no hand-rolled module cache or setInterval).
  * @param rows - Array of objects with a `symbol` property
- * @param cacheKey - Unique key for module-level caching (e.g. 'portfolio', 'watchlist')
+ * @param cacheKey - Distinguishes feeds that share tickers (e.g. 'portfolio', 'watchlist')
  * @param provider - Optional news provider to target (e.g. 'tickertick')
  */
 export function useTickerNews(rows: TickerRow[], cacheKey: string, provider?: string): { items: TickerNewsItem[]; loading: boolean } {
-  const cached = cacheMap.get(cacheKey);
-  const [items, setItems] = useState<TickerNewsItem[]>(() => cached?.items || []);
-  const [loading, setLoading] = useState(!cached);
+  const tickers = (rows || []).map((r) => r.symbol).filter(Boolean);
+  // Sorted so row reordering doesn't churn the query key. The ticker set,
+  // cacheKey, and provider are all in the key, so a change refetches and the
+  // Classic (chain) and Custom ('tickertick') feeds never serve each other's
+  // articles.
+  const tickerKey = [...tickers].sort().join(',');
+  const hasTickers = tickers.length > 0;
 
-  const fetchNews = useCallback(async (): Promise<void> => {
-    const tickers = (rows || []).map((r) => r.symbol).filter(Boolean);
-    const tickerKey = [...tickers].sort().join(',');
-
-    if (!tickers.length) {
-      setItems([]);
-      setLoading(false);
-      cacheMap.set(cacheKey, { items: [], tickerKey: '' });
-      return;
-    }
-
-    setLoading(true);
-    try {
+  const query = useQuery<TickerNewsItem[]>({
+    queryKey: ['dashboard', 'tickerNews', cacheKey, provider ?? null, tickerKey],
+    queryFn: async () => {
       const data = await getNews({ tickers, limit: 50, provider });
-      const mapped: TickerNewsItem[] = data.results?.length > 0 ? mapNewsResults(data.results) : [];
-      setItems(mapped);
-      cacheMap.set(cacheKey, { items: mapped, tickerKey });
-    } catch {
-      setItems([]);
-    } finally {
-      setLoading(false);
-    }
-  }, [rows, cacheKey, provider]);
+      return data.results?.length > 0 ? mapNewsResults(data.results) : [];
+    },
+    enabled: hasTickers,
+    staleTime: NEWS_STALE_MS,
+    refetchInterval: NEWS_POLL_INTERVAL_MS,
+    refetchIntervalInBackground: false,
+  });
 
-  useEffect(() => {
-    const tickers = (rows || []).map((r) => r.symbol).filter(Boolean);
-    const tickerKey = [...tickers].sort().join(',');
-    const cached = cacheMap.get(cacheKey);
-    if (cached?.tickerKey !== tickerKey) {
-      cacheMap.delete(cacheKey);
-    }
-    if (!cacheMap.has(cacheKey)) fetchNews();
-  }, [fetchNews, rows, cacheKey]);
-
-  return { items, loading };
+  return {
+    items: hasTickers ? (query.data ?? []) : [],
+    loading: hasTickers && query.isLoading,
+  };
 }
