+++ web/src/pages/Dashboard/utils/api.ts
@@ -537,8 +537,8 @@ export async function disconnectClaudeOAuth(): Promise<Record<string, unknown>> 
 
 /**
  * Fetch news articles from the native news endpoint.
- * GET /api/v1/news?tickers=...&limit=...&cursor=...
- * @param {{ tickers?: string[], limit?: number, cursor?: string }} opts
+ * GET /api/v1/news?tickers=...&limit=...&cursor=...&provider=...
+ * @param {{ tickers?: string[], limit?: number, cursor?: string, provider?: string }} opts
  * @returns {Promise<{ results: Array, count: number, next_cursor: string|null }>}
  */
 export async function getNews({ tickers, limit = 20, cursor, provider }: NewsParams = {}): Promise<NewsResponse> {
@@ -553,7 +553,9 @@ export async function getNews({ tickers, limit = 20, cursor, provider }: NewsPar
   } catch (e: unknown) {
     const err = e as { message?: string };
     console.error('[API] getNews failed:', err?.message);
-    return { results: [], count: 0, next_cursor: null };
+    // Re-throw so the React Query callers retry (retry:1) and KEEP the last
+    // good list instead of overwriting a live feed with [] on a transient blip.
+    throw e;
   }
 }
